+++ packages/components/checkbox-group/src/index.vue
@@ -1,15 +1,24 @@
 <template>
-   
-    <a-checkbox-group :class="prefixCls" v-bind="$attrs" :options="innerOptions"    @change="onChange"></a-checkbox-group>
+
+  <a-checkbox-group :class="prefixCls" v-bind="$attrs" :options="innerOptions" @change="onChange">
+    <template #label="itemData">
+      {{ itemData[fieldNames.label] }}
+    </template>
+  </a-checkbox-group>
 
 </template>
 
 <script lang="ts" setup>
 
-import type { VzCheckboxGroupProps } from "@viaz/types";
 
+import { useRequest } from "vue-hooks-plus";
 import { useStyle } from "@viaz/hooks";
-import { toRefs, ref } from "vue";
+import { toRefs, ref, watch } from "vue";
+
+import { createNetWork } from "@viaz/utils";
+import type { VzCheckboxGroupProps, DataApi } from "@viaz/types";
+
+import { isFunction, isObject } from "@vue/shared";
 
 const { prefixCls } = useStyle("checkbox-group");
 
@@ -23,16 +32,110 @@ const innerOptions = ref<VzCheckboxGroupProps['options']>([]);
 
 type CheckboxGroupValue = string[] | number[];
 
-const props = withDefaults(defineProps<VzCheckboxGroupProps>(), {});
+const props = withDefaults(defineProps<VzCheckboxGroupProps>(), {
+  fieldNames:{ "label": "name", "value": "id" }
+});
 
-const { options, min, max } = toRefs(props);
 
-(async function initialize() {
-  innerOptions.value = options.value;
+
+const { api, options, params, min, max,fieldNames } = toRefs(props);
+
+
+
+const { data, run } = useRequest(
+  () => {
+    if (isFunction(api)) {
+      return api(params.value);
+    }
+    if (isObject(api)) {
+      const network = createNetWork({}) as any;
+
+      let tmpParams = parseParams();
+
+      let res = network.get({
+        url: generateUrl(api.value as DataApi),
+        params: tmpParams,
+      });
+
+      console.info('res =>',res);
+
+      return res;
+    }
+  },
+  {
+    manual: true,
+  }
+);
+
+const generateUrl = (apiData: DataApi) => {
+  let url = "";
+  if (apiData.protocol !== undefined) {
+    url = `${apiData.protocol}${apiData.hpp}`;
+  } else {
+    url = apiData.hpp as string;
+  }
+
+  return url;
+};
+
+const parseParams = () => {
+  let parsedParams: Record<string, string | number | boolean> = {};
+  params.value?.forEach((item) => {
+    if (item.status === true) {
+      parsedParams[item.key] = item.value;
+    }
+  });
+
+  return parsedParams;
+};
+
+const getOptions = async () => {
+  if (api !== undefined) {
+    run();
+  } else {
+    innerOptions.value = options.value as VzCheckboxGroupProps["options"];
+  }
+};
+
+(async () => {
+  getOptions();
 })();
 
+watch(
+  () => params,
+  () => {
+    getOptions();
+    // run();
+  },
+  {
+    deep: true,
+  }
+);
+
+watch(
+  () => data,
+  () => {
+    console.info('data.value =>',data.value);
+    innerOptions.value = data.value as VzCheckboxGroupProps["options"];
+    // run();
+  },
+  {
+    deep: true,
+  }
+);
+
+watch(
+  () => options,
+  () => {
+    getOptions();
+  },
+  {
+    deep: true,
+  }
+);
+
 const onChange = (checkedValue: CheckboxGroupValue) => {
-  
+
   if (max.value) {
     if (checkedValue.length === max.value) {
       innerOptions.value.forEach((option) => {