+++ packages/components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @viaz/components
 
+## 0.0.2-beta.7
+
+### Patch Changes
+
+- feat checkbox-group
+- Updated dependencies
+  - @viaz/constants@0.0.2-beta.7
+  - @viaz/enums@0.0.2-beta.7
+  - @viaz/hooks@0.0.2-beta.7
+  - @viaz/types@0.0.2-beta.7
+  - @viaz/utils@0.0.2-beta.7
+
 ## 0.0.2-beta.6
 
 ### Patch Changes