+++ packages/directives/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @viaz/directives
 
+## 0.0.2-beta.7
+
+### Patch Changes
+
+- feat checkbox-group
+- Updated dependencies
+  - @viaz/types@0.0.2-beta.7
+
 ## 0.0.2-beta.6
 
 ### Patch Changes